fix(datacell): prevent uint32 multiplication overflow in FlattenDataCell::query (#2188)
In the non-InMemory path of FlattenDataCell::query, the buffer size
calculation id_count * code_size_ was performed in uint32_t arithmetic.
When the product exceeds 2^32, this silently wraps around, allocating a
buffer far smaller than needed and causing heap overflow in subsequent
MultiRead and ScanBatchDists calls.

Cast both operands to uint64_t before multiplication to ensure the
computation is performed in 64-bit arithmetic.

+++ src/datacell/flatten_datacell.h
@@ -325,7 +325,9 @@ FlattenDataCell<QuantTmpl, IOTmpl>::query(float* result_dists,
     }
     if constexpr (not IOTmpl::InMemory) {
         if (id_count > 1) {
-            ByteBuffer codes(id_count * this->code_size_, search_alloc);
+            ByteBuffer codes(
+                static_cast<uint64_t>(id_count) * static_cast<uint64_t>(this->code_size_),
+                search_alloc);
             Vector<uint64_t> sizes(id_count, this->code_size_, search_alloc);
             Vector<uint64_t> offsets(id_count, this->code_size_, search_alloc);
             for (int64_t i = 0; i < id_count; ++i) {
